it seems that the .dll is an interop dll that only works with C#
