+++ src/cameras/archive/nanoscan_32.yml
@@ -7,3 +7,4 @@ dependencies:
     - pip
     - pip:
         - msl-loadlib
+        - pythonnet
